feat: add faq page content
