+++ src/types/shared.types.ts
@@ -1,0 +1,5 @@
+export type FaqItem = {
+  value: string;
+  question: string;
+  answer: React.ReactNode;
+};
